Move arrow_pointer hotspot to better match actual tip
Hotspot position of (6,6) does not fully match the actually tip of the
arrow_pointer cursor image. The position where the click is registered
is visibly off. This can be verified by enabling "Pointer location" in
the developer options.

This patch changes the hotspot to (5,5) which better resembles the
actual tip of the arrow.

Change-Id: Id3c11181c42698c421b90c54e01c1f5f8ec27612